baikks

+++ curso/libft/ft_strrchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pde-jesu <pde-jesu@student.42.fr>          +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 22:49:45 by pde-jesu          #+#    #+#             */
-/*   Updated: 2024/04/22 15:55:16 by pde-jesu         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:53:15 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strrchr(const char *s, int c)
     if (*s == '\0')
 		return NULL;
     i = ft_strlen(s);
-	while (i >= 0)
+	while ((int)i >= 0)
 	{
 		if (s[i] == (char)c)
 			return ((char *)(s + i));
